Update B&A AWS Guide with an alert regarding AWS App Mesh Deprecation (#163)
Update B&A AWS Guide with an alert regarding AWS App Mesh Deprecation Impact.

AWS announced deprecation for App Mesh on September 24, 2024. In response to this announcement, we have made changes to our reference Terraforms to use Load Balancers. We are adding an alert to the relevant section of the AWS Guide to inform the user that they may continue to onboard using Load Balancers.