docs(WIP): Update AWS Guide with notes on Proxy and Perf
Discusses how the root of the differences between AWS and GCP is the difference between AWS Nitro Enclaves and GCP Confidential Compute Spaces, how this requires the AWS TCP-to-VSOCK proxy, and that (and how) this affects server performance.
Also updates the self-serve guide to include our scaling recommendations
for each service, and AWS-specific enclave CPU allocation
recommendations.